moved config to the wrapper file
